fix errors with beans

+++ S3-ITProject/src/main/java/com/fontys/S3ITProject/business/EmployeeService.java
@@ -1,13 +1,14 @@
 package com.fontys.s3itproject.business;
 
-import com.fontys.s3itproject.models.Employee;
+import com.fontys.s3itproject.entity.Employee;
+import org.springframework.stereotype.Service;
 
 import java.util.List;
 
 public interface EmployeeService {
     public boolean createEmployee(Employee employee);
     public List<Employee> readEmployees();
-    public Employee readEmployeeByID(int id);
+    public Employee readEmployeeByID(Long id);
     public boolean updateEmployee(Employee employee);
     public boolean deleteEmployee(Employee employee);
 }
