+++ S3-ITProject/src/main/java/com/fontys/S3ITProject/business/UserService.java
@@ -1,13 +1,14 @@
 package com.fontys.s3itproject.business;
 
-import com.fontys.s3itproject.models.User;
+import com.fontys.s3itproject.entity.User;
+import org.springframework.stereotype.Service;
 
 import java.util.List;
 
 public interface UserService {
     public boolean createUser(User u);
     public List<User> readAllUsers();
-    public User readUserByID(int id);
+    public User readUserByID(Long id);
     public boolean updateUser(User u);
     public boolean deleteUser(User u);
 }
